Process class reifies trips to and from the audio engine.
Processes are state machines, and are passed back and forth with the engine
with ContinueProcess messages.

GetPBHeadProcess is a simple example of using a Process to make a round-trip
reply to a websocket request through the engine.

This included fixing AudioEngine::run() so it actually runs, instead of exiting
after the first go-round.