Updating WAV file sizes on record stop.
Capture actually continues for a couple more buffers after the engine gets the
"stop" message, but meanwhile, the play head is likely to be reset to the start
of the recording, so the read-ahead will want to start reading the new WAV
file(s).  So now we make them valid and correct at that point, preventing an
'invalid-wav-file' error.

Eventually, we'll probably want to update the WAV file size periodically during
recording, which is why the update is called "checkpoint()" here.