Fixing read-ahead.
Basically re-wrote it.  Before, it was working for clips that started at the
play start point, but not so much for clips that started in the middle of
playback.

It's still doing calculations in integer frames relative to the start of the
clip's file.  But these are now 64-bit signed integers -- allowing them to go
negative lets us figure things out more easily, and 64 bits gives enough range
for any conceivable project.